server/requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
black==22.12.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
commonmark==0.9.1
cryptography==39.0.0
dnspython==2.2.1
email-validator==1.3.0
exceptiongroup==1.1.0
fastapi==0.88.0
frozenlist==1.3.3
gunicorn==20.1.0
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
idna==3.4
iniconfig==2.0.0
joblib==1.2.0
loguru==0.6.0
mangum==0.17.0
mongomock==4.1.2
motor==3.1.1
multidict==6.0.4
mypy-extensions==0.4.3
nltk==3.8.1
numpy==1.24.2
packaging==23.0
pandas==2.0.0
pathspec==0.10.3
platformdirs==2.6.2
pluggy==1.0.0
pycparser==2.21
pydantic==1.10.4
Pygments==2.14.0
PyJWT==2.6.0
pymongo==4.3.3
pytest==7.2.0
pytest-asyncio==0.20.3
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2023.3
PyYAML==6.0
regex==2022.10.31
rfc3986==1.5.0
rich==12.6.0
sentinels==1.0.0
shellingham==1.5.0.post1
six==1.16.0
sniffio==1.3.0
starlette==0.22.0
tomli==2.0.1
tqdm==4.64.1
typer==0.7.0
typing-extensions==4.4.0
tzdata==2023.3
uvicorn==0.20.0
watchfiles==0.18.1
websockets==10.4
win32-setctime==1.1.0
yarl==1.8.2
